add transformers to requirements.txt; create tunelite/models/tokenizer.py
