+++ requirements.txt
@@ -1,2 +1,3 @@
 torch
 tqdm
+transformers
